requirements.txt
celery==5.2.3
Flask==2.0.3
Flask-WTF==1.0.0
flower==1.0.0
pytest==7.0.1
redis==4.1.4
virustotal3
psycopg2-binary
Flask-SQLAlchemy
Flask-Hashids
